api/requirements/base.txt
Django
djangorestframework
djangorestframework-simplejwt
mypy 
django-stubs 
djangorestframework-stubs
python-dateutil